+++ smartmin/management/commands/collect_sql.py
@@ -1,18 +1,16 @@
-from __future__ import print_function, unicode_literals
-
-import six
-import sqlparse
 
 from collections import OrderedDict
 from datetime import datetime
-from django.core.management.base import BaseCommand
-from django.db.migrations import RunSQL
-from django.db.migrations.executor import MigrationExecutor
 from enum import Enum
-from six.moves import filter
+
+import sqlparse
 from sqlparse import sql
 from sqlparse import tokens as sql_tokens
 from textwrap import dedent
+
+from django.core.management.base import BaseCommand
+from django.db.migrations import RunSQL
+from django.db.migrations.executor import MigrationExecutor
 
 
 class InvalidSQLException(Exception):
@@ -29,7 +27,6 @@ class SqlType(Enum):
     TRIGGER = 3
 
 
-@six.python_2_unicode_compatible
 class SqlObjectOperation(object):
     def __init__(self, statement, sql_type, obj_name, is_create):
         self.statement = statement
@@ -75,11 +72,15 @@ class SqlObjectOperation(object):
 
         return cls(raw.value.strip(), sql_type, name, is_create)
 
+    def __eq__(self, other):
+        return self.statement == other.statement and self.sql_type == other.sql_type \
+               and self.obj_name == other.obj_name and self.is_create == other.is_create
+
     def __str__(self):
-        return self.statement[:79].replace('\n', ' ')
+        return self.statement[:100].replace('\n', ' ')
 
 
-class Command(BaseCommand):  # pragma: no cover
+class Command(BaseCommand):
     help = """Collects SQL statements from migrations to compile lists of indexes, functions and triggers"""
 
     def add_arguments(self, parser):
@@ -116,7 +117,7 @@ class Command(BaseCommand):  # pragma: no cover
 
         self.write_type_dumps(normalized, preserve_order, output_dir)
 
-    def load_migrations(self):
+    def load_migrations(self):  # pragma: no cover
         """
         Loads all migrations in the order they would be applied to a clean database
         """
@@ -148,7 +149,7 @@ class Command(BaseCommand):  # pragma: no cover
                             operations.append(operation)
 
                             if self.verbosity >= 2:
-                                self.stdout.write(" > %s (%s)" % (operation, migration))
+                                self.stdout.write(" > % -100s (%s)" % (operation, migration))
 
         return operations
 
@@ -159,16 +160,18 @@ class Command(BaseCommand):  # pragma: no cover
         normalized = OrderedDict()
 
         for operation in operations:
-            # do we already have an operation for this object?
-            if operation.obj_name in normalized:
-                if self.verbosity >= 2:
-                    self.stdout.write(" < %s" % normalized[operation.obj_name])
+            op_key = (operation.sql_type, operation.obj_name)
 
-                del normalized[operation.obj_name]
+            # do we already have an operation for this object?
+            if op_key in normalized:
+                if self.verbosity >= 2:
+                    self.stdout.write(" < %s" % normalized[op_key])
+
+                del normalized[op_key]
 
             # don't add DROP operations for objects not previously created
             if operation.is_create:
-                normalized[operation.obj_name] = operation
+                normalized[op_key] = operation
             elif self.verbosity >= 2:
                 self.stdout.write(" < %s" % operation)
 
@@ -194,7 +197,7 @@ class Command(BaseCommand):  # pragma: no cover
         if by_type[SqlType.TRIGGER]:
             self.write_dump('triggers', by_type[SqlType.TRIGGER], output_dir)
 
-    def write_dump(self, type_label, operations, output_dir):
+    def write_dump(self, type_label, operations, output_dir):  # pragma: no cover
         filename = '%s/current_%s.sql' % (output_dir, type_label)
 
         with open(filename, 'w') as f: